pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "meet2task"
version = "0.1.0"
description = "Audio/video to structured text: Groq Whisper, Ollama post-processing, Telegram bot"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [{ name = "dnasonov" }]
keywords = ["telegram", "groq", "whisper", "ollama", "transcription"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
]
dependencies = [
  "groq>=0.4.0",
  "python-telegram-bot>=20.0",
  "python-dotenv>=1.0.0",
  "PyYAML>=6.0",
  "requests>=2.28.0",
  "pydub>=0.25.0",
  "imageio-ffmpeg>=0.4.9",
  "gdown>=5.0.0",
  "beautifulsoup4>=4.9.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[project.scripts]
meet2task-bot = "meet2task.telegram_bot:main"
meet2task-watch = "meet2task.watch_drop:main"
meet2task-transcribe = "meet2task.transcription:main"

[project.urls]
Homepage = "https://github.com/dnasonov/meet2task"
Repository = "https://github.com/dnasonov/meet2task"

[tool.setuptools.packages.find]
where = ["src"]
